# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.4.0
  description: Korean Travel Guide API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    post:
      operationId: createRecommendation
      tags: [recommendations]
      summary: Create and persist a recommendation for the authenticated user
      security: [{ sessionCookie: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRequest"
      responses:
        "201":
          description: Recommendation created
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/RecommendationDataSuccessResponse",
                }
        "200":
          description: Existing idempotent result replayed
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/RecommendationDataSuccessResponse",
                }
        "409":
          description: Idempotency conflict or request in progress
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /recommendations/recent:
    get:
      operationId: getRecentRecommendations
      tags: [recommendations]
      security: [{ sessionCookie: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 20, default: 1 }
      responses:
        "200":
          description: Recent recommendations
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/RecentRecommendationsSuccessResponse",
                }
  /recommendations/{recommendationId}:
    parameters:
      - in: path
        name: recommendationId
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getRecommendation
      tags: [recommendations]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Saved recommendation
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/RecommendationDataSuccessResponse",
                }
    delete:
      operationId: deleteRecommendation
      tags: [recommendations]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Recommendation deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccessBase" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: goat.sid
  schemas:
...
    ApiSuccessBase:
      type: object
      properties:
        success:
          type: boolean
          const: true
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
      required: [success, code, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        requestId:
          type: string
      required: [success, code, message, data]
...
    MoodId:
      type: string
      enum:
        - sea-coast
        - japan-alley
        - alps-ranch
        - forest-garden-rest
        - retro-market-harbor
        - architecture-exhibit-landmark
        - resort-cafe-exotic
    ReferenceCardId:
      type: string
      enum:
        - REF_SEA_01
        - REF_SEA_02
        - REF_SEA_03
        - REF_JP_01
        - REF_JP_02
        - REF_JP_03
        - REF_ALPS_01
        - REF_ALPS_02
        - REF_ALPS_03
        - REF_NATURE_01
        - REF_NATURE_02
        - REF_NATURE_03
        - REF_RETRO_01
        - REF_RETRO_02
        - REF_RETRO_03
        - REF_ARCH_01
        - REF_ARCH_02
        - REF_ARCH_03
        - REF_RESORT_01
        - REF_RESORT_02
        - REF_RESORT_03
...
    TravelPreferences:
      type: object
      properties:
        companion:
          type: string
          enum: [혼자, 연인, 친구, 가족]
        transport:
          type: string
          enum: [자차, 대중교통, 도보중심]
        visitTime:
          type: [string, "null"]
          enum: [새벽, 오전, 한낮, 오후, 일몰, 저녁, 야간, 밤/새벽, null]
        purpose:
          type: string
          enum: [가볍게 산책, 사진 위주, 액티비티, 조용한 휴식]
      required: [companion, transport, purpose]
      additionalProperties: false
    TravelOrigin:
      type: object
      properties:
        type:
          type: string
          enum: [current, region, address, skip]
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        regionName:
          type: string
          minLength: 1
          maxLength: 200
      required: [type]
      additionalProperties: false
...
    CreateRecommendationRequest:
      type: object
      properties:
        moodId: { $ref: "#/components/schemas/MoodId" }
        referenceCardId: { $ref: "#/components/schemas/ReferenceCardId" }
        travelPurpose:
          type: string
          enum: [사진·포토스팟, 산책·힐링, 카페·실내휴식, 전시·건축관람, 체험·액티비티, 먹거리·야간탐방, 숙소·리조트]
        transportType:
          type: string
          enum: [자차, 대중교통, 도보중심]
        visitTime:
          type: string
          enum: [새벽, 오전, 한낮, 오후, 저녁, 야간]
        currentMonth: { type: integer, minimum: 1, maximum: 12 }
        debug: { type: boolean }
        preferences: { $ref: "#/components/schemas/TravelPreferences" }
        origin: { $ref: "#/components/schemas/TravelOrigin" }
        rerollOfRecommendationId: { type: string, format: uuid }
        excludeIds:
          type: array
          uniqueItems: true
          maxItems: 61
          items: { type: string, minLength: 1 }
      oneOf:
        - required: [moodId]
          not:
            required: [referenceCardId]
        - required: [referenceCardId]
          not:
            required: [moodId]
      additionalProperties: false
...
    RecommendationMatchDetail:
      type: object
      properties:
        requested:
          type: array
          items: { type: string }
        matched:
          type: array
          items: { type: string }
        count: { type: integer, minimum: 0 }
        score: { type: number, minimum: 0 }
      required: [requested, matched, count, score]
      additionalProperties: false
    RecommendationThemeScoreDetail:
      type: object
      properties:
        requested: { type: string }
        placeTheme: { type: string }
        matched: { type: boolean }
        score: { type: number, minimum: 0, maximum: 18 }
      required: [placeTheme, matched, score]
      additionalProperties: false
    RecommendationPurposeScoreDetail:
      type: object
      properties:
        requested: { type: string }
        placePurposeTags:
          type: array
          items: { type: string }
        matched: { type: boolean }
        score: { type: number, minimum: 0, maximum: 20 }
      required: [placePurposeTags, matched, score]
      additionalProperties: false
    RecommendationAccessibilityScoreDetail:
      type: object
      properties:
        transportType: { type: string }
        grade: { type: string }
        score: { type: number, minimum: 0, maximum: 12 }
        inferred: { type: boolean }
        note: { type: string }
      required: [score]
      additionalProperties: false
    RecommendationSeasonScoreDetail:
      type: object
      properties:
        requested: { type: string }
        placeSeasonTags:
          type: array
          items: { type: string }
        matchType:
          type: string
          enum: [current, all_season, none, not_requested]
        score: { type: number, minimum: 0, maximum: 13 }
      required: [placeSeasonTags, matchType, score]
      additionalProperties: false
    RecommendationScoreSummary:
      type: object
      description: Typed goat-score-v2 summary. Penalties are stored as non-negative magnitudes and subtracted by the policy.
      properties:
        moodScore: { type: integer, minimum: 0, maximum: 45 }
        conditionScore: { type: integer, minimum: 0, maximum: 45 }
        baseScore: { type: integer, minimum: 0, maximum: 90 }
        originDistanceBonus: { type: integer, minimum: 0, maximum: 10 }
        routeDistanceBonus: { type: integer, minimum: 0, maximum: 10 }
        duplicatePenalty: { type: integer, minimum: 0, maximum: 6 }
        exposurePenalty: { type: integer, minimum: 0, maximum: 5 }
        coverageBoost: { type: integer, minimum: 0, maximum: 3 }
        lowExposureBoost: { type: integer, minimum: 0, maximum: 3 }
        selectionScore: { type: integer, minimum: -11, maximum: 116 }
        displayScore: { type: integer, minimum: 0, maximum: 100 }
      required: [moodScore, conditionScore, baseScore, originDistanceBonus, routeDistanceBonus, duplicatePenalty, exposurePenalty, coverageBoost, lowExposureBoost, selectionScore, displayScore]
      additionalProperties: false
    RecommendationDistanceScoreDetail:
      type: object
      properties:
        distanceKm: { type: [number, "null"], minimum: 0 }
        durationMin: { type: [number, "null"], minimum: 0 }
        source: { type: string, enum: [KAKAO_ROUTE, HAVERSINE, NONE] }
        bonus: { type: integer, minimum: 0, maximum: 10 }
      required: [distanceKm, source, bonus]
      additionalProperties: false
    RecommendationScoreDetails:
      type: object
      properties:
        theme: { $ref: "#/components/schemas/RecommendationThemeScoreDetail" }
        moodTags: { $ref: "#/components/schemas/RecommendationMatchDetail" }
        sceneTags: { $ref: "#/components/schemas/RecommendationMatchDetail" }
        purpose: { $ref: "#/components/schemas/RecommendationPurposeScoreDetail" }
        accessibility: { $ref: "#/components/schemas/RecommendationAccessibilityScoreDetail" }
        season: { $ref: "#/components/schemas/RecommendationSeasonScoreDetail" }
        origin: { $ref: "#/components/schemas/RecommendationDistanceScoreDetail" }
        route: { $ref: "#/components/schemas/RecommendationDistanceScoreDetail" }
      required: [theme, moodTags, sceneTags, purpose, accessibility, season, origin, route]
      additionalProperties: false
...
    RouteInfo:
      type: object
      properties:
        from: { type: string, enum: [ORIGIN, FIRST_CARD] }
        fromLabel: { type: string }
        distanceKm: { type: number, minimum: 0 }
        durationMin: { type: number, minimum: 0 }
        source: { type: string, enum: [KAKAO_ROUTE, HAVERSINE] }
        estimated: { type: boolean }
        scoreApplied: { type: boolean }
      required: [from, fromLabel, distanceKm, source, estimated, scoreApplied]
      additionalProperties: false
...
    RecommendationCardResponse:
      type: object
      properties:
        placeId: { type: string }
        name: { type: string }
        region: { type: string }
        imageUrl: { type: [string, "null"] }
        rank: { type: integer, minimum: 1, maximum: 3 }
        role:
          type: string
          enum: [BEST_SCENE, SAME_MOOD_ALTERNATIVE, CONDITION_FIT_ALTERNATIVE]
        score: { type: integer, minimum: 0, maximum: 100 }
        scoreSummary:
          oneOf:
            - $ref: "#/components/schemas/RecommendationScoreSummary"
            - type: "null"
        scoreDetails:
          oneOf:
            - $ref: "#/components/schemas/RecommendationScoreDetails"
            - type: "null"
        routeInfo:
          oneOf:
            - $ref: "#/components/schemas/RouteInfo"
            - type: "null"
        reason: { type: string }
        reasons:
          type: array
          items: { type: string }
        cautions:
          type: array
          items: { type: string }
        bestSeasons:
          type: array
          items: { type: string }
        seasonBadge:
          type: [object, "null"]
          additionalProperties: true
        crowd:
          $ref: "#/components/schemas/VisitConcentration"
        bookmarked: { type: boolean }
        feedback:
          type: [string, "null"]
          enum: [LIKE, DISLIKE, null]
      required:
        [
          placeId,
          name,
          region,
          imageUrl,
          rank,
          role,
          score,
          scoreSummary,
          scoreDetails,
          routeInfo,
          reason,
          reasons,
          cautions,
          bestSeasons,
          seasonBadge,
          crowd,
          bookmarked,
          feedback,
        ]
    VisitConcentration:
      type: object
      properties:
        level:
          type: string
          enum: [low, medium, high, unknown]
        label: { type: [string, "null"] }
        concentrationRate: { type: [number, "null"] }
        baseDate: { type: [string, "null"] }
        source:
          type: string
          enum: [KTO_VISIT_CONCENTRATION, fallback]
      required: [level, label, concentrationRate, baseDate, source]
    RecommendationData:
      type: object
      properties:
        recommendationId: { type: string, format: uuid }
        policyVersion: { type: string }
        originStatus:
          type: [string, "null"]
          enum: [APPLIED, SKIPPED, UNAVAILABLE, null]
        originNotice: { type: [string, "null"] }
        conditions:
          $ref: "#/components/schemas/CreateRecommendationRequest"
        cards:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: "#/components/schemas/RecommendationCardResponse" }
        course:
          type: [object, "null"]
          additionalProperties: true
        createdAt: { type: string, format: date-time }
      required: [recommendationId, policyVersion, originStatus, originNotice, conditions, cards, course, createdAt]
    RecommendationDataSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/RecommendationData" }
          required: [data]
    RecentRecommendationsSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: "#/components/schemas/RecommendationData" }
                nextCursor: { type: [string, "null"] }
              required: [items, nextCursor]
          required: [data]
...